refactor: major refactoring
